requirements.txt
numpy~=1.23.5
PyYaml~=6.0
requests~=2.28.1